Updated all UTs to pass with changes, new Drop Down Factory yet to be added to UTs.
