+++ src/MockClasses.py
@@ -7,8 +7,9 @@ from typing import Callable, Optional, Any
 import unittest
 from unittest.mock import MagicMock
 
-DEFAULT_GUILD_ID   = 1111111111
-DEFAULT_PROFILE_ID = 1234567890
+DEFAULT_DISPLAY_NAME = "UNIT TESTER 9000"
+DEFAULT_GUILD_ID     = 1111111111
+DEFAULT_PROFILE_ID   = 1234567890
 
 
 #####  Mock Database Interface Class  #####
@@ -195,24 +196,6 @@ class MockDbInterface():
         pass
 
 
-#####  Mock GetPage Class  #####
-
-class MockGetPage():
-
-    async def get_page(self,
-                       index : int):
-        """A bare minimum mock to ensure test compatability.
-
-           Input: self - Pointer to the current object instance.
-                  index - which page to get.
-
-           Output: dis-Embed - a default Eiscord embed object.
-                   int - a valid value of pages for a menu.
-        """
-
-        return dis.Embed(), index
-
-
 #####  Mock Interaction Class  #####
 
 class MockInteraction():
@@ -293,7 +276,8 @@ class MockInteraction():
                Output: none.
             """
 
-            self.id = DEFAULT_PROFILE_ID
+            self.id           = DEFAULT_PROFILE_ID
+            self.display_name = DEFAULT_DISPLAY_NAME
 
     def __init__(self):
         """A bare minimum mock to ensure test compatability.
